removed types2
