+++ models/types.ts
@@ -1,185 +1,159 @@
-// import { Role } from "../constants/role";
-// import { DocumentData, GeoPoint, Timestamp } from "firebase/firestore";
-import { Document, ObjectId, WithId } from "mongodb";
+import { IronSession } from "iron-session";
 
-// TODO: these types may not all be accurate - refer to quiz-real branch when finalizing
+export interface SurveyAnswers {
+    /**
+     * OR relationship
+     * If one category is in the array, show all resources with that category
+     * 
+     * If they send empty array, assume they want everything
+     * 
+     * Your matches: Most relevant based on what you answered
+     * Additional matches: Things you might qualify for but did not ask for
+     */
+        /**
+     * Variable name, not what will be shown on site
+     */
+    categories: string[];
+    householdMembers: number;
+    householdIncome: number;
+    /**
+     * If no answer provided, show all resources regardless of documentation
+     */
+    documentation?: boolean;
+    /**
+     * OR relationship
+     * Use this just for ordering right now
+     */
+        /**
+     * Variable name, not what will be shown on site
+     */
+    languages: string[];
+    /**
+     * same as languages, just for ordering
+     */
+        /**
+     * Variable name, not what will be shown on site
+     */
+    accessibility: string[];
+}
+
+export type IncomeRange = {
+    id: string;
+    minimum: number;
+    maximum: number;
+}
+
 export interface GeoJsonPoint {
     type: "Point",
     coordinates: number[]
 }
+export interface Resource {
+    //Literally the name of the resource
+    name: string;
+    /**
+     * General overview of resource, used in resource page
+     * 
+     * Put other things like eligibility in their proper property
+     */
+    description: string;
+    /**
+     * One sentence used in card description
+     */
+    summary: string;
+    /**
+     * Variable name, not what will be shown on site
+     */
+    category: string[];
+    keywords?: string[];
+    /**
+     * Array where the index + 1 represents the number of household members, and the stored object represents the income range.
+     * 
+     * If there are more household members than the length of the array, use the last range in the array.
+     */
+    incomeByHouseholdMembers?: IncomeRange[];
 
-export interface User {
-    // role: Role;
+    /**
+     * Some resources require some form of GOVERNMENT ISSUED DOCUMENTATION
+     * 
+     * Default to false if not sure
+     */
+    documentationRequired: boolean;
+    headerImageUrl?: string;
+    website: string;
+    /**
+     * Phone and Email represent the communication for the resource in general, hopefully they can redirect to other lines if necessary
+     */
+    phone?: string;
+    email?: string;
+
+    /**
+     * The single location of the main office that manages this program. This is a geopoint we use for the map.
+     */
+    location?: GeoJsonPoint;
+    /**
+     * Similar to location, except it's a string address just used for describing things.
+     */
+    address?: string;
+        /**
+     * Variable name, not what will be shown on site
+     */
+    availableLanguages: string[];
+    /**
+     * this resource fulfills these accessibility options
+     */
+        /**
+     * Variable name, not what will be shown on site
+     */
+    accessibilityOptions?: string[];
+    /**
+     * Description of things you need to be eligible. Separate property so it's easier to format on the page.
+     */
+    /**
+     * put documentation requirements here
+     */
+    eligibilityInfo?: string;
+}
+
+
+export interface Admin {
+    email: string;
+    hashedPassword: string;
     firstName: string;
     lastName: string;
-    email: string;
-    address: Address;
-    birthday: Date;
-    age: number;
-    incomeLevel: number;
+    permissionGroups: PermissionGroup[];
+    title: string;
+    organization: string;
+    phoneNumber: string;
+    profilePicture?: string;
 }
 
-export interface Address {
-    streetNumber: number;
-    streetName: string;
-    city: string;
-    state: string;
-    zip: number;
+export enum PermissionGroup {
+    Admin = "Admin"
 }
 
-export interface FormQuestion {
-    id: number;
-    question: string;
-    description?: string;
+export type ResponseMessage = {
+    message: string
 }
 
-export type UID = string;
 
-// TODO: update this type once Firestore is updated
-// export interface Event extends EventInfo {
-//     id: string;
-// }
-
-// export interface EventInfo {
-//     name: string;
-//     description: string;
-//     summary: string;
-//     location?: GeoPoint;
-//     start?: Timestamp;
-//     end?: Timestamp;
-//     // attendees: UID[];
-// }
-
-export interface SurveyAnswers {
-    category: ResourceCategory[];
-    income?: number;
-    language?: Language[];
-    citizenship?: Citizenship;
-    parentAge?: number;
-    childAge?: number;
-    family?: Family;
-    employmentStatus?: EmploymentStatus;
-    insurance?: Insurance;
-    accessibility?: Accessibility[]; //TODO: Might need to change to multi-select, since people can have multiple accessibility needs
+export type User = {
+    email: string, 
+    isAdmin: boolean
 }
 
-/**
- * deprecated Use ResourceData instead. This should be deleted after ticket #.
- */
-// export interface Service extends DocumentData {
-//     name: string;
-//     description: string;
-//     url: string;
-//     income?: Range[];
-//     language?: String[];
-//     citizenship?: Citizenship;
-//     parentAge?: Range[];
-//     childAge?: Range[];
-//     family?: Family[];
-//     employmentStatus?: EmploymentStatus[];
-//     insurance?: Insurance[];
-//     accessibility?: Accessibility[];
-// }
-export interface Resource {
-    name: string;
-    description?: string; // Markdown
-    category?: ResourceCategory[];
-    minimumIncome?: number;
-    maximumIncome?: number;
-    language?: String[];
-    citizenship?: Citizenship[];
-    minimumParentAge?: number;
-    maximumParentAge?: number;
-    minimumChildAge?: number;
-    maximumChildAge?: number;
-    family?: Family[];
-    employmentStatus?: EmploymentStatus[];
-    insurance?: Insurance[];
-    accessibility?: Accessibility[];
-    location?: GeoJsonPoint;
-    tags?: string[];
-    url?: string;
-    headerImageUrl: string;
-    phoneNumber?: string;
-    email?: string;
-    pointOfContact: string;
-    waitlist?: Waitlist;
-    howToApply?: string; // Markdown
+declare module "iron-session" {
+    interface IronSessionData {
+        user: User
+    }
 }
 
-export type Waitlist = {
-  description: string;
-  helpText?: string;
+export type Invite = {
+    _id: string,
+    expiration: Date
 }
 
-//Inclusive!!!
-export type Range = {
-    minimum: number;
-    maximum: number;
-};
-
-export enum Citizenship {
-    Undocumented = "Undocumented",
-    Non_Immigrant = "Non-Immigrant",
-    Resident="Resident",
-    Citizen="Citizen",
-    Eligible_Non_Citizen="Eligible Non-Citizen"
-}
-
-export enum Family {
-    Nuclear = "Nuclear",
-    Expected = "Expecting",
-    Extended = "Extended",
-    Single_Parent = "Single-Parent",
-    Grandparent = "Grandparent",
-    Foster = "Foster"
-}
-
-export enum EmploymentStatus {
-    Currently_Working = "Currently Working",
-    Seeking_Employment = "Seeking Employment",
-    Studying = "Studying",
-    Stay_At_Home = "Stay at Home"
-}
-
-export enum Insurance {
-    Uninsured = "Uninsured",
-    Private = "Private",
-    Public = "Public"
-}
-
-export enum Accessibility {
-    Speech = "Speech",
-    Vision = "Vision",
-    Hearing = "Hearing",
-    Mental = "Mental",
-    DigitalLiteracy = "Digital Literacy"
-}
-
-export enum Language {
-   English = "English",
-   Spanish = "Spanish",
-   Portuguese = "Portuguese",
-   Chinese = "Chinese",
-   Vietnamese = "Vietnamese",
-   Arabic = "Arabic",
-   Haitian_Creole = "Haitian Creole"
-}
-
-export enum ResourceCategory {
-    Childcare = "Childcare", 
-    Healthcare = "Healthcare", 
-    FinancialHelp = "Financial Help", 
-    ImmigrationAssistance = "Immigration Assistance", 
-    Housing = "Housing"
-}
-
-// TODO: update this type with all sorting methods possible
-// once we know what those are
-export enum ResourceSortingMethod {
-    Alphabetical = "Alphabetical",
-    Sort2 = "Sort2",
-    Sort3 = "Sort3",
-    Sort4 = "Sort4",
-    Sort5 = "Sort5"
+export type ForgotPasswordSession = {
+    _id: string,
+    expiration: Date,
+    email: string
 }
