+++ pages/api/admin/index.tsx
@@ -1,7 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import mongoDbInteractor from "../../../db/mongoDbInteractor";
 import { ObjectId, WithId } from "mongodb";
-import { Admin, ResponseMessage } from "../../../models/types2";
+import { Admin, ResponseMessage } from "../../../models/types";
 import { ADMIN_COLLECTION } from "../../../models/constants";
 import { isInviteValid } from "../../../util/utils";
